refactor: move shared/types/index to shared/types/GetRulesBase/index
https://github.com/akky-xxxx/eslint-plugin-strict-check/issues/285
