+++ src/modules/getRules/index.ts
@@ -1,5 +1,5 @@
-import type { GetRulesBase } from "../../shared/types"
 import type { EsLintRuleBase } from "../../shared/types/EsLintRuleBase"
+import type { GetRulesBase } from "../../shared/types/GetRulesBase"
 import type { RuleName } from "../../shared/types/RuleName"
 
 type GetRules = GetRulesBase<Record<RuleName, EsLintRuleBase>>
