+++ src/shared/types/GetRulesBase/index.ts
@@ -1,3 +1,3 @@
-import type { MasterRecord } from "./MasterRecord"
+import type { MasterRecord } from "../MasterRecord"
 
 export type GetRulesBase<T> = (current: T, master: MasterRecord) => T
